move patch_subclass to __init__
